correction

+++ contracts/CollateralToken.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.5;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract CollateralToken is ERC20 {
-  constructor(uint initialSupply) ERC20("Collateral Token", "CT") {
+  constructor(uint256 initialSupply) ERC20("Collateral Token", "CT") {
     _mint(msg.sender, initialSupply);
   }
 }
